+++ js/settings.js
@@ -36,4 +36,11 @@ export const registerSettings = function () {
 		default: false,
 		type: Boolean
 	});
+
+	game.settings.register(modulename, 'conversion', {
+		scope: "world",
+		config: false,
+		default: false,
+		type: Boolean
+	});
 };