+++ terrain-main.js
@@ -30,6 +30,43 @@ function registerLayer() {
 	});
 }
 
+async function checkUpgrade() {
+	let hasInformed = false;
+	let inform = function () {
+		if (!hasInformed) {
+			ui.notifications.info('Converting old TerrainLayer data, please wait');
+			hasInformed = true;
+		}
+	}
+
+	for (let scene of game.scenes.entries) {
+		if (scene.data.flags?.TerrainLayer) {
+			let gW = scene.data.grid;
+			let gH = scene.data.grid;
+
+			let data = duplicate(scene.data.flags?.TerrainLayer);
+			for (let [k, v] of Object.entries(data)) {
+				if (k == 'costGrid') {
+					let grid = scene.getFlag('TerrainLayer', 'costGrid');
+					for (let y in grid) {
+						for (let x in grid[y]) {
+							//if (Object.values(data).find(t => { return t.x == (parseInt(x) * gW) && t.y == (parseInt(y) * gH); }) == undefined) {
+								inform();
+								let id = makeid();
+								let data = { _id: id, x: parseInt(x) * gW, y: parseInt(y) * gH, points: [[0, 0], [gW, 0], [gW, gH], [0, gH], [0, 0]], width: gW, height: gH, multiple: grid[y][x].multiple };
+								await scene.setFlag('enhanced-terrain-layer', 'terrain' + id, data);
+							//}
+						}
+					}
+				}
+			};
+		}
+	}
+
+	if (hasInformed)
+		ui.notifications.info('TerrainLayer conversion complete.');
+}
+
 export function makeid() {
 	var result = '';
 	var characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
@@ -44,6 +81,13 @@ Hooks.on('canvasInit', () => {
 	canvas.hud.terrain = new TerrainHUD();
 	//Scene.constructor.config.embeddedEntities.Terrain = "terrain";
 });
+
+Hooks.on('ready', () => {
+	if (game.user.isGM && !setting('conversion')) {
+		checkUpgrade();
+		game.settings.set('enhanced-terrain-layer', 'conversion', true);
+	}
+})
 
 Hooks.on('init', () => {
 	game.socket.on('module.enhanced-terrain-layer', async (data) => {